Succession Time Step Bug Fix
Fixed an issue where with certain parameter values the program would attempt to bin multiple keys of the same value in a dictionary. This caused the program to crash and fail. Code now checks to see if the key already exists and if so to bin the value within that key.